move util functions
